requirements.txt
-r requirements-test.txt
attrs==22.1.0
autopep8==2.0.0
dnspython==2.2.1
iniconfig==1.1.1
mccabe==0.7.0
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
tomli==2.0.1
